+++ mine/Traits/MapperTrait.php
@@ -1,10 +1,19 @@
 <?php
+/**
+ * MineAdmin is committed to providing solutions for quickly building web applications
+ * Please view the LICENSE file that was distributed with this source code,
+ * For the full copyright and license information.
+ * Thank you very much for using MineAdmin.
+ *
+ * @Author X.Mo<root@imoi.cn>
+ * @Link   https://gitee.com/xmo/MineAdmin
+ */
 
 namespace Mine\Traits;
 
 use Hyperf\Contract\LengthAwarePaginatorInterface;
 use Hyperf\Database\Model\Builder;
-use Hyperf\Utils\Collection;
+use Hyperf\Database\Model\Model;
 use Mine\Annotation\Transaction;
 use Mine\MineCollection;
 use Mine\MineModel;
@@ -177,9 +186,9 @@ trait MapperTrait
      * @param int $id
      * @return MineModel
      */
-    public function read(int $id): MineModel
+    public function read(int $id): ?MineModel
     {
-        return $this->model::find($id);
+        return ($model = $this->model::find($id)) ? $model : null;
     }
 
     /**
@@ -188,9 +197,9 @@ trait MapperTrait
      * @return MineModel
      * @noinspection PhpUnused
      */
-    public function readByRecycle(int $id): MineModel
+    public function readByRecycle(int $id): ?MineModel
     {
-        return $this->model::withTrashed()->find($id);
+        return ($model = $this->model::withTrashed()->find($id)) ? $model : null;
     }
 
     /**
@@ -213,7 +222,7 @@ trait MapperTrait
     public function update(int $id, array $data): bool
     {
         $this->filterExecuteAttributes($data, true);
-        return $this->model::query()->where((new $this->model)->getKeyName(), $id)->update($data);
+        return $this->model::query()->where((new $this->model)->getKeyName(), $id)->update($data) > 0;
     }
 
     /**
@@ -225,7 +234,9 @@ trait MapperTrait
     {
         foreach ($ids as $id) {
             $model = $this->model::withTrashed()->find($id);
-            $model->forceDelete();
+            if ($model) {
+                $model->forceDelete();
+            }
         }
         return true;
     }
@@ -284,5 +295,74 @@ trait MapperTrait
     public function import(string $dto, ?\Closure $closure = null): bool
     {
         return (new MineCollection())->import($dto, $this->getModel(), $closure);
+    }
+
+    /**
+     * 闭包通用查询设置
+     * @param \Closure|null $closure 传入的闭包查询
+     * @return Builder
+     */
+    public function settingClosure(?\Closure $closure = null): Builder
+    {
+        return $this->model::where(function($query) use($closure) {
+            if ($closure instanceof \Closure) {
+                $closure($query);
+            }
+        });
+    }
+
+    /**
+     * 闭包通用方式查询一条数据
+     * @param \Closure|null $closure
+     * @param array|string[] $column
+     * @return Builder|Model|null
+     */
+    public function one(?\Closure $closure = null, array $column = ['*'])
+    {
+        return $this->settingClosure($closure)->select($column)->first();
+    }
+
+    /**
+     * 闭包通用方式查询数据集合
+     * @param \Closure|null $closure
+     * @param array|string[] $column
+     * @return array
+     */
+    public function get(?\Closure $closure = null, array $column = ['*']): array
+    {
+        return $this->settingClosure($closure)->get($column)->toArray();
+    }
+
+    /**
+     * 闭包通用方式统计
+     * @param \Closure|null $closure
+     * @param string $column
+     * @return int
+     */
+    public function count(?\Closure $closure = null, string $column = '*'): int
+    {
+        return $this->settingClosure($closure)->count($column);
+    }
+
+    /**
+     * 闭包通用方式查询最大值
+     * @param \Closure|null $closure
+     * @param string $column
+     * @return mixed|string|void
+     */
+    public function max(?\Closure $closure = null, string $column = '*')
+    {
+        return $this->settingClosure($closure)->max($column);
+    }
+
+    /**
+     * 闭包通用方式查询最小值
+     * @param \Closure|null $closure
+     * @param string $column
+     * @return mixed|string|void
+     */
+    public function min(?\Closure $closure = null, string $column = '*')
+    {
+        return $this->settingClosure($closure)->min($column);
     }
 }
